restructured project to introduce new architecture
