+++ config/crd/bases/operator.gnmic.dev_targetsources.yaml
@@ -60,6 +60,8 @@ spec:
                         type: string
                     type: object
                 type: object
+              type:
+                type: string
             required:
             - profile
             - provider
