+++ internal/controller/targetsource/core/loaders.go
@@ -1,0 +1,20 @@
+package core
+
+import (
+	"context"
+)
+
+// Loader defines a pluggable TargetSource loader interface
+// Loaders observe external Sources of Truth and emit target snapshots through a channel
+type Loader interface {
+	// Name returns the unique loader identifier e.g. "http_pull"
+	Name() string
+
+	// Start begins discovery and pushes target snapshots into the out channel
+	// The loader must stop cleanly when ctx is cancelled
+	Start(
+		ctx context.Context,
+		targetsourceName string,
+		out chan<- []DiscoveryMessage,
+	) error
+}
