+++ internal/controller/targetsource/core/target_manager.go
@@ -1,4 +1,4 @@
-package targetsource
+package core
 
 import (
 	"context"
